django/backend/requirements.txt
asgiref
async-timeout
attrs
autobahn
Automat
certifi
cffi
channels
channels-redis
charset-normalizer
constantly
cryptography
daphne
Django
django-cors-headers
django-extensions
djangorestframework
djangorestframework-simplejwt
hyperlink
idna
incremental
msgpack
pillow
psycopg2
pyasn1
pyasn1_modules
pycparser
PyJWT
pyOpenSSL
redis
requests
service-identity
setuptools
six
sqlparse
Twisted
txaio
typing_extensions
urllib3
Werkzeug
zope.interface
